Add packages and organize classes.
