+++ app/src/main/java/com/amiel/moviecenter/Authentication/LoginOptions/LoginOptionsFragment.java
@@ -1,4 +1,4 @@
-package com.amiel.moviecenter;
+package com.amiel.moviecenter.Authentication.LoginOptions;
 
 import android.content.Intent;
 import android.graphics.drawable.BitmapDrawable;
@@ -9,19 +9,24 @@ import android.view.View;
 import android.view.ViewGroup;
 import android.widget.Button;
 
+import androidx.activity.result.ActivityResult;
+import androidx.activity.result.ActivityResultCallback;
+import androidx.activity.result.ActivityResultLauncher;
+import androidx.activity.result.contract.ActivityResultContracts;
 import androidx.appcompat.app.AppCompatActivity;
 import androidx.fragment.app.Fragment;
 import androidx.navigation.NavController;
 import androidx.navigation.Navigation;
 
+import com.amiel.moviecenter.Authentication.FirebaseAuthHandler;
 import com.amiel.moviecenter.DB.DatabaseRepository;
 import com.amiel.moviecenter.DB.Model.User;
+import com.amiel.moviecenter.R;
+import com.amiel.moviecenter.Utils.ImageUtils;
 import com.google.android.gms.auth.api.signin.GoogleSignIn;
 import com.google.android.gms.auth.api.signin.GoogleSignInAccount;
 import com.google.android.gms.common.api.ApiException;
 import com.google.android.gms.tasks.Task;
-
-import static com.amiel.moviecenter.FirebaseAuthHandler.RC_SIGN_IN;
 
 public class LoginOptionsFragment extends Fragment {
 
@@ -37,6 +42,7 @@ public class LoginOptionsFragment extends Fragment {
         FirebaseAuthHandler.getInstance().initGoogleAuth(getString(R.string.web_client_id), getActivity());
         db = new DatabaseRepository(getActivity());
         ((AppCompatActivity) getActivity()).getSupportActionBar().hide();
+
         // Defines the xml file for the fragment
         return inflater.inflate(R.layout.login_options_fragment, parent, false);
     }
@@ -49,42 +55,30 @@ public class LoginOptionsFragment extends Fragment {
         signInWithGoogle = (Button) view.findViewById(R.id.login_sign_in_with_google_button);
         signUp = (Button) view.findViewById(R.id.login_sign_up_button);
 
-        signInWithEmail.setOnClickListener(new View.OnClickListener() {
-            public void onClick(View v) {
-                Navigation.findNavController(view).navigate(R.id.signInFragment);
-            }
+        signInWithEmail.setOnClickListener(v -> Navigation.findNavController(view).navigate(R.id.signInFragment));
+
+        signUp.setOnClickListener(v -> Navigation.findNavController(view).navigate(R.id.signUpFragment));
+
+        ActivityResultLauncher<Intent> startActivityIntent = registerForActivityResult(
+                new ActivityResultContracts.StartActivityForResult(),
+                result -> {
+                    Task<GoogleSignInAccount> task = GoogleSignIn.getSignedInAccountFromIntent(result.getData());
+                    try {
+                        GoogleSignInAccount account = task.getResult(ApiException.class);
+
+                        // ID is 0 because were not setting it, it's used just for retrieval
+                        User newUser = new User(account.getDisplayName(), account.getEmail(), ImageUtils.getBytes(((BitmapDrawable)getActivity().getDrawable(R.drawable.default_profile_image)).getBitmap()), 0);
+                        db.insertUserTask(newUser);
+                        NavController navController = Navigation.findNavController(getActivity(), getView().getId());
+                        FirebaseAuthHandler.getInstance().signInWithGoogle(account, getActivity(), navController);
+                    } catch (ApiException e) {
+                        Log.w("TAG", "Google sign in failed", e);
+                    }
+                });
+
+        signInWithGoogle.setOnClickListener(v -> {
+            Intent signInIntent = FirebaseAuthHandler.getInstance().getSignInIntent();
+            startActivityIntent.launch(signInIntent);
         });
-
-        signUp.setOnClickListener(new View.OnClickListener() {
-            public void onClick(View v) {
-                Navigation.findNavController(view).navigate(R.id.signUpFragment);
-            }
-        });
-
-        signInWithGoogle.setOnClickListener(new View.OnClickListener() {
-            public void onClick(View v) {
-                Intent signInIntent = FirebaseAuthHandler.getInstance().getSignInIntent();
-                startActivityForResult(signInIntent, RC_SIGN_IN);
-            }
-        });
-    }
-
-    @Override
-    public void onActivityResult(int requestCode, int resultCode, Intent data) {
-        super.onActivityResult(requestCode, resultCode, data);
-        if (requestCode == RC_SIGN_IN) {
-            Task<GoogleSignInAccount> task = GoogleSignIn.getSignedInAccountFromIntent(data);
-            try {
-                GoogleSignInAccount account = task.getResult(ApiException.class);
-
-                // ID is 0 because were not setting it, it's used just for retrieval
-                User newUser = new User(account.getDisplayName(), account.getEmail(), ImageUtils.getBytes(((BitmapDrawable)getActivity().getDrawable(R.drawable.default_profile_image)).getBitmap()), 0);
-                db.insertUserTask(newUser);
-                NavController navController = Navigation.findNavController(getActivity(), getView().getId());
-                FirebaseAuthHandler.getInstance().signInWithGoogle(account, getActivity(), navController);
-            } catch (ApiException e) {
-                Log.w("TAG", "Google sign in failed", e);
-            }
-        }
     }
 }
