+++ app/src/main/java/com/amiel/moviecenter/Authentication/SignIn/SignInFragment.java
@@ -1,4 +1,4 @@
-package com.amiel.moviecenter;
+package com.amiel.moviecenter.Authentication.SignIn;
 
 import android.os.Bundle;
 import android.util.Patterns;
@@ -13,6 +13,9 @@ import androidx.fragment.app.Fragment;
 import androidx.navigation.NavController;
 import androidx.navigation.Navigation;
 
+import com.amiel.moviecenter.Authentication.FirebaseAuthHandler;
+import com.amiel.moviecenter.R;
+import com.amiel.moviecenter.Utils.TextValidator;
 import com.google.android.material.textfield.TextInputEditText;
 import com.google.android.material.textfield.TextInputLayout;
